chore: moved date exported method to internal package
