+++ internal/date/date.go
@@ -4,6 +4,8 @@ import (
 	"github.com/influxdata/flux/codes"
 	"github.com/influxdata/flux/internal/errors"
 	"github.com/influxdata/flux/internal/zoneinfo"
+	"github.com/influxdata/flux/interpreter"
+	"github.com/influxdata/flux/semantic"
 	"github.com/influxdata/flux/values"
 )
 
@@ -24,4 +26,42 @@ func GetTimeInLocation(t values.Time, location string, offset values.Duration) (
 		return values.NewTime(timeWithOffset), nil
 	}
 	return values.NewTime(t), nil
+}
+
+func GetLocationFromObjArgs(args values.Object) (string, values.Duration, error) {
+	a := interpreter.NewArguments(args)
+	return GetLocationFromFluxArgs(a)
+}
+
+func GetLocationFromFluxArgs(args interpreter.Arguments) (string, values.Duration, error) {
+	location, err := args.GetRequiredObject("location")
+	if err != nil {
+		return "UTC", values.ConvertDurationNsecs(0), err
+	}
+	return GetLocation(location)
+}
+
+func GetLocation(location values.Object) (string, values.Duration, error) {
+	var (
+		name, offset values.Value
+		ok           bool
+	)
+
+	name, ok = location.Get("zone")
+	if !ok {
+		return "UTC", values.ConvertDurationNsecs(0), errors.New(codes.Invalid, "zone property missing from location record")
+	} else if got := name.Type().Nature(); got != semantic.String {
+		return "UTC", values.ConvertDurationNsecs(0), errors.Newf(codes.Invalid, "zone property for location must be of type %s, got %s", semantic.String, got)
+	}
+
+	if offset, ok = location.Get("offset"); ok {
+		if got := offset.Type().Nature(); got != semantic.Duration {
+			return "UTC", values.ConvertDurationNsecs(0), errors.Newf(codes.Invalid, "offset property for location must be of type %s, got %s", semantic.Duration, got)
+		}
+	}
+
+	if name.IsNull() {
+		return "UTC", offset.Duration(), nil
+	}
+	return name.Str(), offset.Duration(), nil
 }
