+++ libflux/go/libflux/buildinfo.gen.go
@@ -597,7 +597,7 @@ var sourceHashes = map[string]string{
 	"stdlib/universe/union_heterogeneous_test.flux":                                               "5158d9efdf88ec10945e84f72cb7ed9669b25f967b7b4f778d2456d52ecd34da",
 	"stdlib/universe/union_test.flux":                                                             "8529f2f609d9876f975ad648dec24eb60b569e8ac1c38e4f980e2234e68149a9",
 	"stdlib/universe/unique_test.flux":                                                            "02180f651906f48cc57cc777d443a64bd2155394ac0ade68888f47f5900e5408",
-	"stdlib/universe/universe.flux":                                                               "4dee405c283abb4d8c88a8daf5da190cfa163cc801c9568f4e85494e728c3ee6",
+	"stdlib/universe/universe.flux":                                                               "c8a3bbd18143113965ca29f94e8ff86c68eea21427fb1e971b4b8334ff2f8802",
 	"stdlib/universe/universe_truncateTimeColumn_test.flux":                                       "7ca30c57336b22ae430a082d33589fa919f4558c3ed6dbe37ecde367d602435c",
 	"stdlib/universe/window_aggregate_test.flux":                                                  "e2dba18647b2a180cafe81dd00a43077eb489421d229785e63b270eb79ddb2b1",
 	"stdlib/universe/window_default_start_align_test.flux":                                        "162f9d452fed411edfdcb11ef1b8ae0c565d87bbc351596d277c7db55fe67e8f",
